requirements.txt
Babel==2.3.4
Django==1.9.6
Pillow==3.2.0
Unidecode==0.04.19
django-appconf==1.0.2
django-compressor==2.0
django-debug-toolbar==1.4
django-extensions==1.6.7
django-extra-views==0.7.1
django-haystack==2.4.1
django-tables2==1.2.1
django-treebeard==4.0.1
django-widget-tweaks==1.4.1
phonenumbers==7.3.2
psycopg2==2.6.1
purl==1.3
pytz==2016.4
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
sorl-thumbnail==12.3
sqlparse==0.1.19
